backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
PyJWT==2.10.1
python-dotenv==1.0.1
supabase==2.15.0
